pyproject.toml
[project]
name = "ollama-deproxy"
version = "0.1.0"
description = "For allow local usage ollama behind some proxy aka OpenWebUI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.132.0",
    "httpx[brotli,http2,zstd]>=0.28.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.41.0",
]
classifiers = ["Private :: Do Not Upload"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
# If you use src/ layout (recommended)
packages = ["src/ollama_deproxy"]

[project.scripts]
ollama-deproxy = "ollama_deproxy.main:run"

[tool.hatch.build.sdist]
relative-install = true


# Code-style tools

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.flake8]
max-line-length = 120

[dependency-groups]
dev = [
    "black>=26.1.0",
    "ruff>=0.15.2",
]

